Crate voting_tools_rs

Rust implementation of voting tools

Original Haskell repository is https://github.com/input-output-hk/voting-tools

The queries themselves (as well as the details of the CLI) are different, but they should produce similar outputs. Malformed registrations are silently ignored.

Re-exports

pub use crate::verification::*;

Modules

cli 🔒

data 🔒

db 🔒

error 🔒

exports 🔒

logic 🔒

test_api

Test api for internal tests as well for building mocks and various testing utils outside the project

testing 🔒

verification

Structs

Args

CLI arguments for snapshot tool

DbConfig

Information required to connect to a database

InvalidRegistration

A registration that failed validation, along with the error

RegistrationCorruptedBin

Registrations with a corrupted raw cbor binary require extra metadata for context

Sig

An ED25519 signature

Signature

A signature for a registration as defined in CIP-15

SlotNo

A slot number

SnapshotEntry

Single element in a snapshot

VotingPowerArgs

Arguments to the voting_power function

VotingPurpose

A u64 used to identify the purpose of a particular registration

Enums

DryRunCommand

Sub command for internal testing or dry runs

RegistrationError

An error encountered during parsing and validation of a Catalyst registration

VotingKey

The source of voting power for a given registration

Functions

show_error_warning

If there are errors, we want to notify the user, but it’s not really actionable, so we provide the option to silence the error via env var

voting_power

Calculate voting power info by querying a db-sync instance