Module voting_tools_rs::data

Re-exports

pub use crypto2::PubKey;

pub use crypto2::Sig;

pub use network_id::NetworkId;

Modules

crypto2 ğŸ”’

hash ğŸ”’

hex_bytes

network_id ğŸ”’

Macros

hex_impls ğŸ”’

Structs

DbHost

Database host

DbName

Database name

DbPass

Database password

DbUser

Database user

Nonce

RawRegistration

A Raw Registration from the DB.arbitrary_signed_registration

Registration

A catalyst registration on Cardano in either CIP-15 or CIP-36 format

RewardsAddress

A rewards address in a catalyst registation

Signature

A signature for a registration as defined in CIP-15

SignedRegistration

A Catalyst registration in either CIP-15 or CIP-36 format, along with its signature

SlotNo

A slot number

SnapshotEntry

Single element in a snapshot

StakeAddr

StakeKeyHex

StakePubKey

TxId

VotingKeyHex

VotingPurpose

A u64 used to identify the purpose of a particular registration

__WrapperDbPass ğŸ”’

Enums

VotingKey

The source of voting power for a given registration

Constants

DELEGATIONS_OR_DIRECT ğŸ”’

KEY_61284 ğŸ”’

KEY_61285 ğŸ”’

NETWORK_ID ğŸ”’

NONCE ğŸ”’

PAYMENT_ADDRESS ğŸ”’

STAKE_ADDRESS ğŸ”’

VOTE_KEY ğŸ”’

VOTE_PURPOSE ğŸ”’

WEIGHT ğŸ”’

WITNESS ğŸ”’

Functions

_strategy_of_0 ğŸ”’

_strategy_of_0 ğŸ”’

_strategy_of_0 ğŸ”’

_strategy_of_0 ğŸ”’

_strategy_of_0 ğŸ”’

_strategy_of_0 ğŸ”’

_to_fn_ptr ğŸ”’

_to_fn_ptr ğŸ”’

_to_fn_ptr ğŸ”’

_to_fn_ptr ğŸ”’

_to_fn_ptr ğŸ”’

_to_fn_ptr ğŸ”’

inspect_cip36_reg ğŸ”’

inspect_cip36_sig ğŸ”’

inspect_metamap_reg ğŸ”’

inspect_metamap_sig ğŸ”’

inspect_nonce ğŸ”’

inspect_rewards_addr ğŸ”’

inspect_stake_key ğŸ”’

inspect_voting_key ğŸ”’

inspect_voting_purpose ğŸ”’

inspect_witness ğŸ”’

To produce the witness field in the case of a staking public key, the CBOR representation of a map containing a single entry with key 61284 and the registration metadata map is signed with the staking key as follows: first, the blake2b-256 hash of the data is obtained. This hash is then signed using the Ed25519 signature algorithm. The witness metadata entry is added to the transaction under key 61285. This function extracts and verifies the witness signature provided in the registration.

ox_hex ğŸ”’

ox_hex_ ğŸ”’

ox_hex_sig ğŸ”’

Opaque Types

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’

ğŸ”’