Module voting_tools_rs::error

Structs

InvalidRegistration

A registration that failed validation, along with the error

RegistrationCorruptedBin

Registrations with a corrupted raw cbor binary require extra metadata for context

Enums

RegistrationError

An error encountered during parsing and validation of a Catalyst registration