Function voting_tools_rs::voting_power

pub fn voting_power(
    db: &dyn DataProvider,
    min_slot: Option<SlotNo>,
    max_slot: Option<SlotNo>,
    testnet_magic: Option<TestnetMagic>
) -> Result<Vec<Output>>

Calculate voting power info by querying a db-sync instance

Invalid registrations are silently ignored (e.g. if they contain bad/null JSON metadata, if they have invalid signatures, etc).

If provided, min_slot and max_slot can be used to constrain the time period to query. If None they default to:

min_slot: 0

max_slot: i64::MAX

Together they form an inclusive range (i.e. blocks with values equal to min_slot or max_slot are included)

Errors

Returns an error if either of lower or upper doesn’t fit in an i64