Trait voting_tools_rs::DataProvider

pub trait DataProvider: Debug {
    fn vote_registrations(
        &self,
        lower: Option<SlotNo>,
        upper: Option<SlotNo>
    ) -> Result<Vec<Reg>>;
    fn stake_values<'a>(
        &self,
        stake_addrs: &'a [String]
    ) -> Result<HashMap<&'a str, BigDecimal>>;
}

Abstraction trait over data provider for voting tools. This approach can allow various data sources for registration queries like standard db sync database or in memory one including mocks.

Required Methods

fn vote_registrations(
&self,
lower: Option<SlotNo>,
upper: Option<SlotNo>
) -> Result<Vec<Reg>>

Retrieves voter registration for optional bounds. They need to be expressed as absolute slot numbers

Errors

Returns error on reading data issue

fn stake_values<'a>(
&self,
stake_addrs: &'a [String]
) -> Result<HashMap<&'a str, BigDecimal>>

Retrieves stakes values for given array of addresses

Errors

Returns error on reading data issue