Function voting_tools_rs::verification::verify::is_valid_rewards_address

pub fn is_valid_rewards_address(
    rewards_address_prefix: &u8,
    network: NetworkId
) -> bool

Reward addresses start with a single header byte identifying their type and the network, followed by 28 bytes of payload identifying either a stake key hash or a script hash. Function accepts this first header prefix byte. Validates first nibble is within the address range: 0x0? - 0x7? + 0xE? , 0xF? Validates second nibble matches network id: 0/1