Crate voting_tools_rs

Rust implementation of voting tools

Original Haskell repository is https://github.com/input-output-hk/voting-tools

The queries themselves (as well as the details of the CLI) are different, but they should produce similar outputs. Malformed registrations are silently ignored.

Modules

test_api

Test api for internal tests as well for building mocks and various testing utils outside the project

vectors

Test vectors for various CIPs

Structs

Args

CLI arguments for snapshot tool

Db

A handle to the db-sync database instance

DbConfig

Information required to connect to a database

Sig

An ED25519 signature

Signature

A signature for a registration as defined in CIP-15

SlotNo

A slot number

SnapshotEntry

Single element in a snapshot

VotingPowerArgs

Arguments to the voting_power function

VotingPurpose

A u64 used to identify the purpose of a particular registration

Enums

DryRunCommand

Sub command for internal testing or dry runs

VotingPowerSource

The source of voting power for a given registration

Traits

DataProvider

Abstraction trait over data provider for voting tools. This approach can allow various data sources for registration queries like standard db sync database or in memory one including mocks.

Functions

show_error_warning

If there are errors, we want to notify the user, but it’s not really actionable, so we provide the option to silence the error via env var

voting_power

Calculate voting power info by querying a db-sync instance

Type Definitions

Conn

Type alias for the connection type provided to diesel code