Trait voting_tools_rs::DataProvider

pub trait DataProvider: Debug {
    fn vote_registrations(
        &self,
        lower: SlotNo,
        upper: SlotNo
    ) -> Result<Vec<SignedRegistration>>;
    fn stake_values(
        &self,
        stake_addrs: &[StakeKeyHex]
    ) -> Result<HashMap<StakeKeyHex, BigDecimal>>;
}

Abstraction trait over data provider for voting tools. This approach can allow various data sources for registration queries like standard db sync database or in memory one including mocks.

Required Methods

fn vote_registrations(
&self,
lower: SlotNo,
upper: SlotNo
) -> Result<Vec<SignedRegistration>>

Get all vote registrations and signatures between two slot numbers

If either slot number is None, they are ignored

Errors

This function returns an error if a database error occurs. The exact details of any error will depend on the database implementation

fn stake_values(
&self,
stake_addrs: &[StakeKeyHex]
) -> Result<HashMap<StakeKeyHex, BigDecimal>>

Retrieves stakes values for given array of addresses

Errors

This function returns an error if a database error occurs. The exact details of any error will depend on the database implementation